requirements.txt
pymupdf
Pillow
pydantic
google-genai
python-dotenv
python-docx
